Major reStructure, closes #13. Fill comments on classes needed
