+++ app/DAL/dal.service.ts
@@ -100,23 +100,3 @@ export class NanaDal {
         return Observable.throw(error.json().error || 'Server error');
     }
 }
-
-/**
- * (description)
- * 
- * @export
- * @class MockNanaDal
- */
-export class MockNanaDal {
-    /**
-     * (description)
-     * 
-     * @returns {Observable<Object>} (description)
-     */
-    getItems(): Observable<Object> {
-        return Observable.of({
-            NavigateID: 1,
-            NavigateName: 'name1'
-        });
-    }
-}
