+++ app/main.ts
@@ -1,6 +1,4 @@
 import {bootstrap}    from '@angular/platform-browser-dynamic';
 import {AppComponent} from './app.component';
-import {NanaDalComponent} from './nanaDataService/DAL/nana.dal.component';
 
-//bootstrap(NanaDalComponent);
 bootstrap(AppComponent);
