--- app/nanaDataService/navigationService/navigation.service.test.ts
@@ -1,14 +1,0 @@
-import {describe, expect, beforeEachProviders, it, inject } from '@angular/core/testing';
-//import { NanaserviceService} from './nanaservice.service';
-
-describe('Service  NanaserviceService', () => {
-
-    it('has name', () => {
-        let i: number;
-        i = 1;
-        expect(i).toEqual(1);
-    });
-})
-
-
-
