+++ testjasmine.html
@@ -30,8 +30,8 @@
 
     // #3. Import the spec file explicitly
    Promise.all([
-    System.import('app/nanaDataService/nanaserviceService/nanaservice.service.test'),
-    System.import('app/nanaDataService/DAL/nana.dal.test')
+    System.import('app/DAL/dal.service.spec'),
+    System.import('app/Navigation/navigation.service.spec')
   ])
